+++ context/app/static/js/api/scfind/useCellTypeCountForTissue.ts
@@ -1,5 +1,5 @@
 import useSWR from 'swr';
-import { fetcher } from 'js/helpers/swr';
+import { fetcher, multiFetcher } from 'js/helpers/swr';
 import { useAppContext } from 'js/components/Contexts';
 import { createScFindKey } from './utils';
 
@@ -16,12 +16,15 @@ export interface CellTypeCountForTissueParams {
   tissue: string;
 }
 
-type CellTypeCountForTissueKey = string;
+type CellTypeCountForTissueKey = string | null;
 
 export function createCellTypeCountForTissueKey(
   scFindEndpoint: string,
   { tissue }: CellTypeCountForTissueParams,
-): CellTypeCountForTissueKey {
+): CellTypeCountForTissueKey | null {
+  if (!tissue || tissue.length === 0) {
+    return null;
+  }
   return createScFindKey(scFindEndpoint, 'cellTypeCountForTissue', {
     tissue,
   });
@@ -31,4 +34,15 @@ export default function useCellTypeCountForTissue(props: CellTypeCountForTissueP
   const { scFindEndpoint } = useAppContext();
   const key = createCellTypeCountForTissueKey(scFindEndpoint, props);
   return useSWR<CellTypeCountsForTissue, unknown, CellTypeCountForTissueKey>(key, (url) => fetcher({ url }));
+}
+
+export function useCellTypeCountForTissues(tissues: string[]) {
+  const { scFindEndpoint } = useAppContext();
+  const keys = tissues
+    .map((tissue) => createCellTypeCountForTissueKey(scFindEndpoint, { tissue }))
+    .filter((key): key is string => key !== null);
+
+  return useSWR<CellTypeCountsForTissue[], unknown, NonNullable<CellTypeCountForTissueKey>[]>(keys, (urls: string[]) =>
+    multiFetcher({ urls }),
+  );
 }