+++ context/app/static/js/components/detailPage/PublicationsSection/index.ts
@@ -1,0 +1,3 @@
+import PublicationsSection from './PublicationsSection';
+
+export default PublicationsSection;